packages/python/requirements.txt
fastapi>=0.115.0
PyJWT[crypto]>=2.8.0
